cleanup PageBulkExportService
